Remove the need for AtomicI64

+++ src/hal/dummy/time.rs
@@ -8,7 +8,7 @@
 //
 
 #[inline]
-pub fn now_us() -> i64 {
+pub fn now_us() -> u32 {
     0
 }
